Extract common functionality in superclass
